Took out the creation of the hist file from inside the if statement.
This was necessary because ROOT wouldn't load the macro due to there being a reference to the hist_file outside of the if statement. It seems ROOT couldn't find it being defined inside of the if statement. Also, it turns out it isn't necessary, the "UPDATE" option already takes into account the possibility of the file not existing yet and creates it. You learn something every day.