Created "CompareSimulation.C" file.
The goal of this macro is, for a given run, compare the actual data with simulated ER only and NR only data. Through this comparison, we wish to obtain the NR to ER ratio that would best match with the actual data.
Also, I'm attempting to use the TDataFrame class, which seems to allow for some interesting manipulation of data. This should be a valuable learning experience.