Wrote function "Createf90Dist" to obtain the desired f90 distributions.
The function takes a number of arguments:
"file_name" = the name of the .root file where the desired data is.
"charge_max" = the upper limit on the charge being considered.
"f90_max" = the upper limit on f90 to be considered when constructing the histogram.
"f90_min" = the lower limit on f90 to be considered when constructing the histogram.
"division_size" = the size of the charge selection which will be used to divide the total charge in divisions of equal size.
"bin_number" = more appropriate to call it "division_number". It determines witch division we wish to construct the histogram from, thus defining the upper and lower limits of the total charge to consider.