Createf90DIst -> made it so the generated histogram is returned
Was previously causing a error because I was attempting to close the TFile from witch the data wa extracted. However, since all generated histograms automatically go into the current directory, which is the by default the currently opened file, closing the file resulting in the histogram being unaccessible. Fixed this problem with the command "SetDirectory(0)", which makes it so the histogram is no longer located in the current directory, therefore allowing me to still access the histogram even after the file has already been closed.