Generated the remaining two histograms,  also created a THStack.
The THStack will hold the three histograms (data,  ER only,  NR only) and then write them all in the same plot. This is only for visualization.
The remaining two histograms are generated from the data, but each one containing only one of the event peaks: one for ER and one for NR. The limits for these histograms are defined by the user, though the default value should separate the peaks well enough. These same range values are used when generating histograms from the ER and NR only simulation data, which should allow for a clear comparison between the two.