Change the "CompareSimulation" macro.
Included an extra function,  "CompareAllHist". It does basically everything that was previously directly written in the macro. I turned it into a function because generating all the histograms is time consuming and they may be moments where we only wish to test changes in the comparison algorithm. As such, the new macro has an option to indicate whether the histograms to be compared already exist or not. In the case they do, the macro takes them and runs the comparison algorithm (yet to be written). If not, the macro first calls the "GenerateAllHist" function and only then runs the algorithm.