Changed recoil energy ranges for next simulations
The main goal is to obtain a resulting maximum charge value for the distribution of 2000 PE, for both the ER and NR case. Once this is done, we separate the distribution in bins of 20PE up to 1000PE. For each bin, obtain the f90 distribution and calculate its RMS. Do the same for the actual data. Finally, plot RMS as a function of charge for all cases.