Created 'PlotRMS.C' file.
This file will contain the macro used to obtain the plot of RMS as a function of total charge. The RMS will be obtained from f90 histograms generated for charge bins of size 20 PE between 0 and 1000 PE. For each bin, a f90 histogram will be generated from a actual data, an ER simulation and a NR simulation.