Changed file name from 'PlotRMS.C' to 'CreateF90Histograms.C'
The reason is I've decide to write to separate macros: one that generates all desired histograms, and another which will access those histograms, obtain their RMS and plot them. 'CreateF90Histograms.C' creates the histograms and seems to be running smoothly. All that's left is to write the second macro.