Changed the max energy once again.
The reason to do so is the same as the last time. The last simulation resulted in a considerable excess of the max total charge, going up to 2000 PE. This means I can reduce the energy further and still get at least events of 1000 PE. So that's what I'm doing now.