Wrote GraphDiff() function and plotted the relative differences.
The function mentioned takes two graphs and returns a graph that is the difference of the two. This function was used to compute the relative difference between the data f90 RMS and MC simulation f90 RMS. The plots obtained are then saved in the corresponding files.
TO DO: Make a plot containing the relative differences of the ER and NR together.