Wrote 'GenerateF90vChargePlot()' function.
This function substitutes the need for the prior three distinct plot functions (GenerateMeanF90vChargePlot(), GeneratePeakF90vChargePlot() and GenerateRMSF90vChargePlot()') as it can generate the outcome of each one by itself. This was done by simply adding a new parameter, stupidly called 'parameter'  witch indicates witch of the plots (Mean, Peak or RMS v Charge Total) to create.
I now realize I could've achieved the same effect by writing a function that just called one of the three prior functions, but I think having one for each seems to be a waste of code, as the differences are minimal and have only to do with file and plot naming.