Create "ToF.C" file.
The purpose of this macro is to analyse the distribution of the time of flight parameter. This parameter doesn't already exist in the reconstructed root file of the run, and must be calculated. I'm unsure about what is the best way to define it, but for now have chosen to use the start_time of the SIDeltaE detector and the start time of the S1 signal.