Added "DefineCuts()" function.
This function defines some of the cuts to be used when generating the f90 histograms. Currently, it applies quality cuts, such as number of clusters and rep as well as ToF cuts. Future improvements may be separating the two as 'quality' and 'boundary' cuts and adding other possible boundary restrictions (such as charge and f90).