Changed the parameters 'GenerateF90Hist()' expects
It now takes only two parameters:
TString file_name >> the name of the file were the data from witch the histogram will be constructed.
TCut histogram_cuts >> all the cuts to be applied when constructing the histograms. Previously, the function expected certain ranges and from these values it would construct the cuts itself. However, since it was recently changed to also expect a TCut objects containing quality cuts, it made sense to remove the parameters that would define the ranges. The user must now provide these ranges within the cut that is given to the function. WriteF90Hist() was also changed to match.