Rewrote the macro.
It was temporary, so I simply used what was already there to write something a bit more complex. This code takes a given ToF range, divides it into a number of bins of equal size and for each bin plots a f90 distribution. The goal is to see how the distribution changes with f90. A further study might involve computing what is the percentage of event with the ToF range that would be classified as ER.