Finished writing the code to compare run sets.
In this case,  the runs being compared are 1489 up to 1521, with some exceptions stated by the code. The code creates a number of histograms of different parameters (S1, S2, drift time and f90) to qualitatively check if the runs are similar to each other, such that merging them for further analysis makes sense.