Renamed file: GroupAnalysis.C -> LowBeToF.C
Also wrote a function that can possibly be used to define a TCutG object to be used when selecting the low energy Beriloum blob. The function writes a contour accepting regions with at lest 10 events. The largest contour is selected (since the distribution isn't well behave, it results in a number of very small contours that are irrelevant) and saved as a TGraph object. This TGraph object can later be called and converted in a TCutG object to be used when selecting the low Be blob, at least in theory. Testing is still pending.