Wrote code to select a "Electron Recoil Region".
The code computes the 3% and 97% quantile of a f90 histogram for a given s1 charge bin. Each point is then saved in two different graphs such as to define a region that classifies an event as being an electron recoil. Ideally, one would do this by using Am runs and others that have almost exclusively ER events, though this is not currently possible. A temporary remedy would be to do this with Li beam runs by only selecting events in a given f90 range, though this would certainly add some unwanted bias, especially in the region were ER and NR events overlap.