Modified "Bananator()" and added "LowBeGraphCut()"
Bananator -> the modifications were simply to stop the function from drawing a histogram whenever it is called.
LowBeGraphCut -> This function takes the graph saved in 'LowBeCut.root' and returns it as a TCutG object that can be used to select the low energy berilium blob in the deltaE by E histogram.