Wrote "SigmaEllipse" function.
This was wrote in an attempt to use the fit parameters obtained from a bigaussian fit to the deltaE vs E histogram to construct a ellipse. It has proven to be successful. The TGraph object obtained from this process will from know on be used to select the low energy berilium blob, substituting the previous method.